add update admin done

+++ Clinnic/src/main/java/com/owen/repository/AppointmentRepository.java
@@ -58,6 +58,6 @@ public interface AppointmentRepository {
     
     List<Appointment> getAppointmentsbyIDPrefordelete(int id);
     
-    Integer CountAppointmentbyUser(User u);
+    long CountAppointmentbyUser(User u);
 
 }
